Move resolution code for ConcreteUpdateStatement
